Big refactor. Move Init() into config, fix errors
ClipseConfig's HistoryFilePath and TempDirPath now hold the full paths
to the respective files. This obsoleted functions like Paths() and
the practice of having historyFilePath be passed into functions. Left in
for compatibility, but it would be best if those were removed.

HistoryFilePath and TempDirPath both take relative or absolute paths,
including just the file name (will be treated as relative). From a user
perspective, no breaking changes. Just more configurability.

I won't remove obsolete code mentioned above as that will change a lot
of the codebase. So I will leave that to @savedra1.
